+++ bncwebservices/src/com/acc/services/CustomerHealthDataService.java
@@ -1,5 +1,5 @@
 /**
- * 
+ *
  */
 package com.acc.services;
 
@@ -8,12 +8,12 @@ import com.accenture.model.CustomerHealthDataModel;
 
 /**
  * @author swarnima.gupta
- * 
+ *
  */
 public interface CustomerHealthDataService
 {
 	/**
-	 * 
+	 *
 	 * @param customerId
 	 * @param heartBeatRate
 	 * @param bloodPressure
@@ -24,10 +24,10 @@ public interface CustomerHealthDataService
 	 * @return String
 	 */
 	public String saveCustomerHealthData(final String customerId, final Integer heartBeatRate, final Integer bloodPressure,
-			final Integer milesRun, final Integer caloriesBurned, final Integer timeTaken, final Integer age);
+			final Double milesRun, final Integer caloriesBurned, final Double timeTaken, final Integer age);
 
 	/**
-	 * 
+	 *
 	 * @param customerId
 	 * @return CustomerHealthDataModel
 	 */
